# lib/api-zod/src/generated/types/tutorAskInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TutorAskInput {
  /** @nullable */
  sessionId?: number | null;
  message: string;
  /** @nullable */
  selectedLectureText?: string | null;
  /** @nullable */
  lectureId?: number | null;
  /** @nullable */
  topicId?: number | null;
  /** @nullable */
  problemId?: number | null;
}